Move request_schema to request
